+++ src/server.rs
@@ -3,3 +3,7 @@ pub mod tcp;
 
 #[cfg(feature = "tcp")]
 pub mod tcp_server;
+
+mod service;
+
+pub use service::{NewService, Service};
